+++ tests/BackEnd/TcaTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use OliverKlee\Realty\BackEnd\Tca;
+
 /**
  * Test case.
  *
@@ -8,7 +10,7 @@
 class tx_realty_BackEnd_TcaTest extends \Tx_Phpunit_TestCase
 {
     /**
-     * @var tx_realty_Tca
+     * @var Tca
      */
     private $fixture = null;
 
@@ -20,7 +22,7 @@ class tx_realty_BackEnd_TcaTest extends \Tx_Phpunit_TestCase
     protected function setUp()
     {
         $this->testingFramework = new Tx_Oelib_TestingFramework('tx_realty');
-        $this->fixture = new tx_realty_Tca();
+        $this->fixture = new Tca();
     }
 
     protected function tearDown()